feat: Add quest routes
